requirements.txt
mcp[cli]>=1.0.0
asyncpg>=0.29.0
uvicorn>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
